+++ src/petuum_ps/thread/ps_msgs.hpp
@@ -713,7 +713,7 @@ protected:
 
 struct ServerRowRequestReplyMsg : public ArbitrarySizedMsg {
 public:
-  explicit ServerRowRequestReplyMsg(int32_t avai_size) {
+  explicit ServerRowRequestReplyMsg(size_t avai_size) {
     own_mem_ = true;
     mem_.Alloc(get_header_size() + avai_size);
     InitMsg(avai_size);
@@ -764,7 +764,7 @@ public:
   }
 
 protected:
-  virtual void InitMsg(int32_t avai_size) {
+  virtual void InitMsg(size_t avai_size) {
     ArbitrarySizedMsg::InitMsg(avai_size);
     get_msg_type() = kServerRowRequestReply;
   }
@@ -772,7 +772,7 @@ protected:
 
 struct ClientSendOpLogMsg : public ArbitrarySizedMsg {
 public:
-  explicit ClientSendOpLogMsg(int32_t avai_size) {
+  explicit ClientSendOpLogMsg(size_t avai_size) {
     own_mem_ = true;
     mem_.Alloc(get_header_size() + avai_size);
     InitMsg(avai_size);
@@ -818,7 +818,7 @@ public:
   }
 
 protected:
-  virtual void InitMsg(int32_t avai_size) {
+  virtual void InitMsg(size_t avai_size) {
     ArbitrarySizedMsg::InitMsg(avai_size);
     get_msg_type() = kClientSendOpLog;
   }
@@ -826,7 +826,7 @@ protected:
 
 struct ServerPushRowMsg : public ArbitrarySizedMsg {
 public:
-  explicit ServerPushRowMsg(int32_t avai_size) {
+  explicit ServerPushRowMsg(size_t avai_size) {
     own_mem_ = true;
     mem_.Alloc(get_header_size() + avai_size);
     InitMsg(avai_size);
@@ -866,7 +866,7 @@ public:
   }
 
 protected:
-  virtual void InitMsg(int32_t avai_size) {
+  virtual void InitMsg(size_t avai_size) {
     ArbitrarySizedMsg::InitMsg(avai_size);
     get_msg_type() = kServerPushRow;
   }
